+++ hutool-core/src/main/java/cn/hutool/core/lang/tree/Tree.java
@@ -315,22 +315,25 @@ public class Tree<T> extends LinkedHashMap<String, Object> implements Node<T> {
 	 */
 	public Tree<T> cloneTree() {
 		final Tree<T> result = ObjectUtil.clone(this);
-		result.setChildren(cloneChildren());
+		result.setChildren(cloneChildren(result));
 		return result;
 	}
 
 	/**
 	 * 递归复制子节点
 	 *
+	 * @param parent 新的父节点
 	 * @return 新的子节点列表
 	 */
-	private List<Tree<T>> cloneChildren() {
+	private List<Tree<T>> cloneChildren(final Tree<T> parent) {
 		final List<Tree<T>> children = getChildren();
 		if (null == children) {
 			return null;
 		}
 		final List<Tree<T>> newChildren = new ArrayList<>(children.size());
-		children.forEach((t) -> newChildren.add(t.cloneTree()));
+		children.forEach((t) -> {
+			newChildren.add(t.cloneTree().setParent(parent));
+		});
 		return newChildren;
 	}